refactor to use dto package instead of model for openApi models
