+++ src/main/java/it/gov/pagopa/payhub/activities/mapper/NotificationQueueMapper.java
@@ -1,7 +1,7 @@
 package it.gov.pagopa.payhub.activities.mapper;
 
 import it.gov.pagopa.payhub.activities.dto.debtposition.DebtPositionDTO;
-import it.gov.pagopa.pu.p4paionotification.model.generated.NotificationQueueDTO;
+import it.gov.pagopa.pu.p4paionotification.dto.generated.NotificationQueueDTO;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
