+++ src/test/java/it/gov/pagopa/payhub/activities/activity/debtposition/ionotification/SendDebtPositionIONotificationActivityTest.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.payhub.activities.activity.debtposition.ionotification;
 import it.gov.pagopa.payhub.activities.dto.debtposition.DebtPositionDTO;
 import it.gov.pagopa.payhub.activities.mapper.NotificationQueueMapper;
 import it.gov.pagopa.payhub.activities.service.debtposition.ionotification.SendDebtPositionIONotificationService;
-import it.gov.pagopa.pu.p4paionotification.model.generated.NotificationQueueDTO;
+import it.gov.pagopa.pu.p4paionotification.dto.generated.NotificationQueueDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
